ER/TD model refactor clean_text()
