+++ libs/osomerank/topic_diversity.py
@@ -3,7 +3,7 @@
 """
 Created on Sun Jun 16 03:45:55 2024
 
-@author: saumya
+@author: saumya (modified by Bao)
 """
 
 from bertopic import BERTopic
@@ -13,6 +13,7 @@ import traceback
 
 import os
 import configparser
+from osomerank.utils import clean_text
 
 TD_MEAN = 2.66
 TD_STD = 1.77
@@ -35,6 +36,7 @@ with open(
 ) as ff:
     topic_diversity = json.load(ff)
 
+
 def remove_urls(text, replacement_text=""):
     # Define a regex pattern to match URLs
     url_pattern = re.compile(r"https?://\S+|www\.\S+")
@@ -45,32 +47,12 @@ def remove_urls(text, replacement_text=""):
     return text_without_urls
 
 
-# Filtering text: remove special characters, alphanumeric, return None if text doesn't meet some criterial like just one word or something
-def process_text(text):
-    text_urls_removed = remove_urls(text)
-    if len(text.strip().split(" ")) <= 3:
-        return "NA"
-    return text_urls_removed
-
-
-# Filtering text: remove special characters, alphanumeric, return None if text doesn't meet some criterial like just one word or something
-def process_text_multiple(texts):
-    processed_texts = []
-    for text in texts:
-        text_urls_removed = remove_urls(text)
-        if len(text_urls_removed.strip().split(" ")) <= 3:
-            processed_texts.append("NA")
-        else:
-            processed_texts.append(text_urls_removed)
-    return processed_texts
-
-
 def td_prediction(feed_posts, sm_type):
-    
+
     audience_diversity_val = [-1000] * len(feed_posts)
 
     try:
-        
+
         sm_texts = []
         text_index = []
 
@@ -92,7 +74,7 @@ def td_prediction(feed_posts, sm_type):
                         sm_texts.append(feed_post["text"])
                         text_index.append(i)
 
-        sm_texts_processed = process_text_multiple(sm_texts)
+        sm_texts_processed = [clean_text(text) for text in sm_texts]
         texts_processed = []
         text_index_processed = []
 
@@ -105,20 +87,21 @@ def td_prediction(feed_posts, sm_type):
 
         for i in range(len(topics)):
             if int(topics[i]) != -1:
-                td_val = topic_diversity[
-                    str(topics[i])
-                ]
-                audience_diversity_val[text_index_processed[i]] = (td_val - TD_MEAN)/TD_STD
+                td_val = topic_diversity[str(topics[i])]
+                audience_diversity_val[text_index_processed[i]] = (
+                    td_val - TD_MEAN
+                ) / TD_STD
 
     except Exception:
         print(traceback.format_exc())
-        return [mean_topic_diversity]*len(feed_posts)
+        return [mean_topic_diversity] * len(feed_posts)
 
     for i in range(len(audience_diversity_val)):
         if audience_diversity_val[i] == -1000:
             audience_diversity_val[i] = mean_topic_diversity
 
     return audience_diversity_val
+
 
 def ad_prediction_single(feed_post, sm_type):
     audience_diversity_val = -1000
@@ -134,12 +117,12 @@ def ad_prediction_single(feed_post, sm_type):
                 sm_text = feed_post["text"]
         else:
             sm_text = feed_post["text"]
-        if process_text(sm_text) != "NA":
-            sm_text = process_text(sm_text)
+        if clean_text(sm_text) != "NA":
+            sm_text = clean_text(sm_text)
             topic = BERTopic_model_loaded.transform([sm_text])[0][0]
             if int(topic) != -1:
                 audience_diversity_val = topic_diversity[str(topic)]
-                
+
     except Exception as e:
         print(e)
         return mean_topic_diversity
